src configured
