+++ menu-public-portal/src/common/helpers/slug.helpers.ts
@@ -1,0 +1,2 @@
+export const formatSlugFromUrlParam = (param: string): string =>
+  Boolean(param) ? `/${param}` : "";
