+++ menu-public-portal/src/pages/index.tsx
@@ -1,0 +1,29 @@
+import type { NextPage } from "next";
+import Link from "next/link";
+import Head from "next/head";
+import Image from "next/image";
+import styles from "../styles/Home.module.css";
+
+const Home: NextPage = () => {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Tu Carta - Restaurantes</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <main className={styles.main}>
+        <h5 className={styles.title}>Listado restaurantes</h5>
+
+        <Link href="/restaurante/papulinos">Papulinos</Link>
+      </main>
+
+      <footer className={styles.footer}>
+        <p>Lemoncode</p>
+      </footer>
+    </div>
+  );
+};
+
+export default Home;
